+++ trunk/ServerBase/common_config.cpp
@@ -111,7 +111,7 @@ ServerConfigItem* CommonConfig::GetServerConfigItem(const TCHAR* strServerName)
 	return NULL;
 }
 
-ServerConfigItem* CommonConfig::GetServerConfigItemById(uint16 iServerId)
+ServerConfigItem* CommonConfig::GetServerConfigItemById(uint8 iServerId)
 {
 	std::map<std::wstring, ServerConfigItem>::iterator mit = m_mServerConfigItems.begin();
 	while (mit != m_mServerConfigItems.end())
